requirements.txt
treys>=0.1.8
pyyaml>=6.0
